+++ src/Controller/Admin/CustomDashboardCrudController.php
@@ -37,7 +37,7 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
     /**
      * Redirect /admin to the reports page to hide easyAdmin's default front page.
      */
-    #[Route('/admin')]
+    #[Route('/admin', name: 'admin')]
     public function index(): Response
     {
         return $this->redirectToRoute('dashboard', ['entityType' => 'report']);
@@ -62,17 +62,25 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
             $formData = [];
         }
 
-        $formParameters = array_merge([
-            'groups' => [],
-            'subowner' => '',
-            'theme' => '',
-            'category' => '',
-            'self_service' => '',
-            'search' => '',
-        ], $formData);
+        $formParameters = array_merge(
+            [
+                'groups' => [],
+                'subowner' => '',
+                'theme' => '',
+                'category' => '',
+                'self_service' => '',
+                'search' => '',
+            ],
+            $formData
+        );
 
-        $userGroups = $this->entityManager->getRepository(UserGroup::class)->findAll();
-        $userGroupsThemesAndCategories = $this->getUserGroupsThemesAndCategories($userGroups ?: [], $entityType);
+        $userGroups = $this->entityManager
+            ->getRepository(UserGroup::class)
+            ->findAll();
+        $userGroupsThemesAndCategories = $this->getUserGroupsThemesAndCategories(
+            $userGroups ?: [],
+            $entityType
+        );
 
         // Get a query for the entity type.
         $repository = $this->getRepository($entityType);
@@ -80,7 +88,10 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
         $query = $this->applyFilters($query, $formParameters, $entityType);
 
         // Get sub owners if a group has been selected.
-        $subOwnerOptions = $this->getSubOwnerOptions($repository, $formParameters['groups']);
+        $subOwnerOptions = $this->getSubOwnerOptions(
+            $repository,
+            $formParameters['groups']
+        );
 
         $paginator = $this->paginator->paginate(
             $query,
@@ -92,15 +103,20 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
         $themes = [];
 
         if (!empty($formParameters['groups'])) {
-            $groups = $this->entityManager->getRepository(UserGroup::class)->findBy([
-                'id' => $formParameters['groups'],
-            ]);
+            $groups = $this->entityManager
+                ->getRepository(UserGroup::class)
+                ->findBy([
+                    'id' => $formParameters['groups'],
+                ]);
         } else {
             $groups = $userGroups;
         }
 
         foreach ($groups as $group) {
-            $groupThemes = 'report' == $entityType ? $group->getReportThemes() : $group->getSystemThemes();
+            $groupThemes =
+                'report' == $entityType
+                    ? $group->getReportThemes()
+                    : $group->getSystemThemes();
 
             foreach ($groupThemes as $theme) {
                 if ('' != $formParameters['theme']) {
@@ -114,7 +130,10 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
 
                     foreach ($theme->getOrderedCategories() as $category) {
                         if ('' != $formParameters['category']) {
-                            if ($category->getId() != $formParameters['category']) {
+                            if (
+                                $category->getId() !=
+                                $formParameters['category']
+                            ) {
                                 continue;
                             }
                         }
@@ -127,8 +146,19 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
 
         $selfServiceOptions = $this->getSelfServiceOptions($entityType);
 
-        $filterFormBuilder = $this->getFilterFormBuilder($userGroupsThemesAndCategories, $formParameters, $subOwnerOptions, true, true, $selfServiceOptions);
-        $filterFormBuilder->setMethod('GET')->setAction($this->generateUrl('dashboard', ['entityType' => $entityType]));
+        $filterFormBuilder = $this->getFilterFormBuilder(
+            $userGroupsThemesAndCategories,
+            $formParameters,
+            $subOwnerOptions,
+            true,
+            true,
+            $selfServiceOptions
+        );
+        $filterFormBuilder
+            ->setMethod('GET')
+            ->setAction(
+                $this->generateUrl('dashboard', ['entityType' => $entityType])
+            );
 
         return $this->render('dashboard.html.twig', [
             'paginator' => $paginator,
@@ -146,28 +176,38 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
      *
      * @throws \Exception
      */
-    private function getUserGroupsThemesAndCategories(array $userGroups, string $entityType): mixed
-    {
-        return array_reduce($userGroups,
+    private function getUserGroupsThemesAndCategories(
+        array $userGroups,
+        string $entityType,
+    ): mixed {
+        return array_reduce(
+            $userGroups,
             function ($carry, UserGroup $group) use ($entityType) {
                 $carry['groups'][$group->getId()] = $group->getName();
 
-                $groupThemes = 'report' == $entityType ? $group->getReportThemes() : $group->getSystemThemes();
+                $groupThemes =
+                    'report' == $entityType
+                        ? $group->getReportThemes()
+                        : $group->getSystemThemes();
 
                 foreach ($groupThemes as $theme) {
                     $carry['themes'][$theme->getId()] = $theme->getName();
 
                     foreach ($theme->getOrderedCategories() as $category) {
-                        $carry['categories'][$category->getId()] = $category->getName();
+                        $carry['categories'][
+                            $category->getId()
+                        ] = $category->getName();
                     }
                 }
 
                 return $carry;
-            }, [
+            },
+            [
                 'groups' => [],
                 'themes' => [],
                 'categories' => [],
-            ]);
+            ]
+        );
     }
 
     /**
@@ -177,8 +217,9 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
      *
      * @return ReportRepository|SystemRepository|null
      */
-    private function getRepository(string $entityType): ReportRepository|SystemRepository|null
-    {
+    private function getRepository(
+        string $entityType,
+    ): ReportRepository|SystemRepository|null {
         return match ($entityType) {
             'system' => $this->entityManager->getRepository(System::class),
             'report' => $this->entityManager->getRepository(Report::class),
@@ -195,7 +236,10 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
      *
      * @return QueryBuilder
      */
-    private function applyFilters(QueryBuilder $query, array $formParameters, ?string $entityType = null,
+    private function applyFilters(
+        QueryBuilder $query,
+        array $formParameters,
+        ?string $entityType = null,
     ): QueryBuilder {
         $query->andWhere('e.archivedAt IS NULL');
 
@@ -208,32 +252,51 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
 
         // Get the groups the user can search in.
         if (!empty($formParameters['groups'])) {
-            $groups = $this->entityManager->getRepository(UserGroup::class)->findBy([
-                'id' => $formParameters['groups'],
-            ]);
+            $groups = $this->entityManager
+                ->getRepository(UserGroup::class)
+                ->findBy([
+                    'id' => $formParameters['groups'],
+                ]);
 
             foreach ($groups as $group) {
-                $query->andWhere($query->expr()->isMemberOf(':group'.$group->getId(), 'e.groups'));
+                $query->andWhere(
+                    $query
+                        ->expr()
+                        ->isMemberOf(':group'.$group->getId(), 'e.groups')
+                );
                 $query->setParameter(':group'.$group->getId(), $group);
             }
         }
 
-        if (isset($formParameters['self_service']) && '' != $formParameters['self_service']) {
-            $item = $this->entityManager->getRepository(SelfServiceAvailableFromItem::class)->findOneBy([
-                'id' => $formParameters['self_service'],
-            ]);
+        if (
+            isset($formParameters['self_service'])
+            && '' != $formParameters['self_service']
+        ) {
+            $item = $this->entityManager
+                ->getRepository(SelfServiceAvailableFromItem::class)
+                ->findOneBy([
+                    'id' => $formParameters['self_service'],
+                ]);
             if (null != $item) {
-                $query->andWhere(':self_service MEMBER OF e.selfServiceAvailableFromItems');
+                $query->andWhere(
+                    ':self_service MEMBER OF e.selfServiceAvailableFromItems'
+                );
                 $query->setParameter('self_service', $item);
             }
         }
 
-        if (isset($formParameters['search']) && '' != $formParameters['search']) {
+        if (
+            isset($formParameters['search'])
+            && '' != $formParameters['search']
+        ) {
             $query->andWhere('e.name LIKE :name');
             $query->setParameter('name', '%'.$formParameters['search'].'%');
         }
 
-        if (isset($formParameters['subowner']) && '' != $formParameters['subowner']) {
+        if (
+            isset($formParameters['subowner'])
+            && '' != $formParameters['subowner']
+        ) {
             $query->andWhere('e.sysOwnerSub = :subowner');
             $query->setParameter('subowner', $formParameters['subowner']);
         }
@@ -244,11 +307,15 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
     /**
      * Get sub-owners for selected group.
      */
-    private function getSubOwnerOptions(mixed $repository, mixed $selectedGroups): mixed
-    {
-        $groups = $this->entityManager->getRepository(UserGroup::class)->findBy([
-            'id' => $selectedGroups,
-        ]);
+    private function getSubOwnerOptions(
+        mixed $repository,
+        mixed $selectedGroups,
+    ): mixed {
+        $groups = $this->entityManager
+            ->getRepository(UserGroup::class)
+            ->findBy([
+                'id' => $selectedGroups,
+            ]);
 
         $subOwnersQueryBuilder = $repository->createQueryBuilder('e');
         $subOwnersQueryBuilder->select('DISTINCT e.sysOwnerSub');
@@ -261,22 +328,34 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
         if (Report::class === $class) {
             $subOwnersQueryBuilder->andWhere('e.sysStatus = \'Aktiv\'');
         } elseif (System::class === $class) {
-            $subOwnersQueryBuilder->andWhere('e.sysStatus <> \'Systemet bruges ikke længere\'');
+            $subOwnersQueryBuilder->andWhere(
+                'e.sysStatus <> \'Systemet bruges ikke længere\''
+            );
         }
 
         foreach ($groups as $group) {
-            $subOwnersQueryBuilder->andWhere($subOwnersQueryBuilder->expr()->isMemberOf(':group'.$group->getId(), 'e.groups'));
-            $subOwnersQueryBuilder->setParameter(':group'.$group->getId(), $group);
+            $subOwnersQueryBuilder->andWhere(
+                $subOwnersQueryBuilder
+                    ->expr()
+                    ->isMemberOf(':group'.$group->getId(), 'e.groups')
+            );
+            $subOwnersQueryBuilder->setParameter(
+                ':group'.$group->getId(),
+                $group
+            );
         }
 
         $subOwners = $subOwnersQueryBuilder->getQuery()->getResult();
 
-        return array_reduce($subOwners,
+        return array_reduce(
+            $subOwners,
             function ($carry, $item) {
                 $carry[$item['sysOwnerSub']] = $item['sysOwnerSub'];
 
                 return $carry;
-            }, []);
+            },
+            []
+        );
     }
 
     /**
@@ -291,7 +370,9 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
         $selfServiceOptions = [];
         if ('system' == $entityType) {
             /* @var Collection $selfServiceAvailableFromItems */
-            $selfServiceAvailableFromItems = $this->entityManager->getRepository(SelfServiceAvailableFromItem::class)->findAll();
+            $selfServiceAvailableFromItems = $this->entityManager
+                ->getRepository(SelfServiceAvailableFromItem::class)
+                ->findAll();
             /* @var SelfServiceAvailableFromItem $item */
             foreach ($selfServiceAvailableFromItems as $item) {
                 $selfServiceOptions[$item->getName()] = $item->getId();
@@ -329,7 +410,9 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
             'multiple' => true,
             'attr' => [
                 'class' => 'form-control',
-                'data-placeholder' => $this->translator->trans('filter.placeholder.groups'),
+                'data-placeholder' => $this->translator->trans(
+                    'filter.placeholder.groups'
+                ),
             ],
             'required' => false,
             'data' => $formParameters['groups'] ?? null,
@@ -349,7 +432,9 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
             $filterFormBuilder->add('theme', ChoiceType::class, [
                 'label' => 'filter.theme',
                 'placeholder' => 'filter.placeholder.theme',
-                'choices' => array_flip($userGroupsThemesAndCategories['themes']),
+                'choices' => array_flip(
+                    $userGroupsThemesAndCategories['themes']
+                ),
                 'attr' => [
                     'class' => 'form-control',
                 ],
@@ -361,7 +446,9 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
             $filterFormBuilder->add('category', ChoiceType::class, [
                 'label' => 'filter.category',
                 'placeholder' => 'filter.placeholder.category',
-                'choices' => array_flip($userGroupsThemesAndCategories['categories']),
+                'choices' => array_flip(
+                    $userGroupsThemesAndCategories['categories']
+                ),
                 'attr' => [
                     'class' => 'form-control',
                 ],
@@ -405,14 +492,22 @@ class CustomDashboardCrudController extends AbstractSystatusDashboardController
     public function showAction(): RedirectResponse|Response
     {
         $entityArray = $this->entity;
-        if (Report::class == $entityArray['class'] || System::class == $entityArray['class']) {
+        if (
+            Report::class == $entityArray['class']
+            || System::class == $entityArray['class']
+        ) {
             $entity = $this->getEntity($entityArray['class'], $_GET['id']);
             $accessGranted = $this->isGranted('show', $entity);
 
             if (!$accessGranted) {
-                $this->addFlash('error', $this->translator->trans('flash.access_denied'));
+                $this->addFlash(
+                    'error',
+                    $this->translator->trans('flash.access_denied')
+                );
 
-                return $this->redirectToRoute('list', ['entityType' => strtolower($entityArray['name'])]);
+                return $this->redirectToRoute('list', [
+                    'entityType' => strtolower($entityArray['name']),
+                ]);
             }
         }